requirements.txt
SQLAlchemy==2.0.36
pandas==2.2.3
python-dotenv==1.0.1
typer==0.12.5
tabulate==0.9.0
streamlit==1.38.0
matplotlib==3.9.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
dateparser==1.2.0
